feat: Issue #476 Phase 3 logger.py分割とDiscord通知無効化
## logger.py大規模リファクタリング完了
- 1,806行 → 70行 (97%削減)
- Single Responsibility Principleによる4ファイル分割:
  * logging_formatters.py - フォーマッター処理専用
  * logging_handlers.py - メインロガー管理
  * performance_logger.py - パフォーマンス監視とエラー分析
  * structured_logger.py - 構造化ログとClaude Code連携

## Discord通知完全無効化
- 3つのDiscord通知スクリプトを無効化
- WebhookURL削除とcurl通信コメントアウト
- 実行権限削除で安全性確保

## 技術的負債解消効果
- コード保守性大幅向上
- 機能別責任分離による可読性改善
- 後方互換性完全保持
- mypy strict mode完全対応
